Hoist internal packages; remove util
