+++ 2024/day01/day01_test.go
