+++ 2024/day01/example_file.txt
