+++ 2024/day01/parser.go
