+++ 2024/day01/parser_test.go
