+++ 2024/day02/day02.go
