+++ 2024/day02/day02_test.go
