+++ 2024/day02/example_file.txt
