+++ 2024/day02/parser.go
