+++ 2024/day02/parser_test.go
