+++ 2024/day03/day03.go
