+++ 2024/day03/day03_test.go
