--- 2024/internal/util/int.go
@@ -1,9 +1,0 @@
-package util
-
-// Absolute value for int
-func IntAbs(n int) int {
-	if n < 0 {
-		return -n
-	}
-	return n
-}
