+++ 2024/main.go
@@ -5,7 +5,3 @@ import "github.com/cpmachado/advent-of-code/2024/cmd"
 func main() {
 	cmd.Execute()
 }
-
-
-
-
